+++ tests/test_validators.py
@@ -7,7 +7,7 @@ from __future__ import absolute_import, division, print_function
 import pytest
 import zope.interface
 
-from attr.validators import instance_of, provides
+from attr.validators import instance_of, provides, optional
 from attr._compat import TYPE
 from . import simple_attr
 
@@ -107,4 +107,75 @@ class TestProvides(object):
         assert (
             "<provides validator for interface {interface!r}>"
             .format(interface=IFoo)
+        ) == repr(v)
+
+
+class TestOptional(object):
+    """
+    Tests for `optional`.
+    """
+    def test_success_with_type(self):
+        """
+        Nothing happens if types match.
+        """
+        v = optional(instance_of(int))
+        v(None, simple_attr("test"), 42)
+
+    def test_success_with_none(self):
+        """
+        Nothing happens if None.
+        """
+        v = optional(instance_of(int))
+        v(None, simple_attr("test"), None)
+
+    def test_subclass(self):
+        """
+        Subclasses are accepted too.
+        """
+        v = optional(instance_of(int))
+        v(None, simple_attr("test"), True)
+
+    def test_given_interface(self):
+        """
+        Nothing happens if value provides requested interface.
+        """
+        @zope.interface.implementer(IFoo)
+        class C(object):
+            def f(self):
+                pass
+
+        v = optional(provides(IFoo))
+        v(None, simple_attr("x"), C())
+
+    def test_no_interface(self):
+        """
+        Nothing happens if None passed for an interface too.
+        """
+        v = optional(provides(IFoo))
+        v(None, simple_attr("x"), None)
+
+    def test_fail(self):
+        """
+        Raises `TypeError` on wrong types.
+        """
+        v = optional(instance_of(int))
+        a = simple_attr("test")
+        with pytest.raises(TypeError) as e:
+            v(None, a, "42")
+        assert (
+            "'test' must be <{type} 'int'> (got '42' that is a <{type} "
+            "'str'>).".format(type=TYPE),
+            a, int, "42",
+
+        ) == e.value.args
+
+    def test_repr(self):
+        """
+        Returned validator has a useful `__repr__`.
+        """
+        v = optional(instance_of(int))
+        assert (
+            ("<optional validator for <instance_of validator for type "
+             "<{type} 'int'>> or None>")
+            .format(type=TYPE)
         ) == repr(v)